requirements.txt
customtkinter==5.2.1
Pillow==10.1.0
reportlab==4.0.4
openpyxl==3.1.2